refactor: api 함수 경로 수정, .env 파일 적용
