--- src/constants/url.js
@@ -1,2 +1,0 @@
-export const BASE_URL = 'http://idol-service.com';
-
